moved ProgressBar into misc package, added preview and refactored imports where necessary
