+++ app/src/main/java/com/studhub/app/presentation/auth/components/SignInWithGoogle.kt
@@ -5,7 +5,7 @@ import androidx.compose.runtime.LaunchedEffect
 import androidx.hilt.navigation.compose.hiltViewModel
 import com.studhub.app.core.utils.ApiResponse
 import com.studhub.app.presentation.auth.AuthViewModel
-import com.studhub.app.presentation.ui.common.ProgressBar
+import com.studhub.app.presentation.ui.common.misc.ProgressBar
 
 @Composable
 fun SignInWithGoogle(
