+++ app/src/main/java/com/studhub/app/presentation/ui/common/misc/ProgressBar.kt
@@ -1,4 +1,4 @@
-package com.studhub.app.presentation.ui.common
+package com.studhub.app.presentation.ui.common.misc
 
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.fillMaxSize
